Update UnifiNetworkMaps.ps1
You don't need to define array first. When you have $Variable = foreach ... it will create it and put all items into it. There's one thing to know thou. If you don't add the [Array] before, if it will be  just 1 element this will be a string type. If you want it to behave like an array it's important to add the type before it. I've modified two examples giving you two examples.